Переносит фильтрацию в селекторы
